requirements.txt
gradio>=4.0.0
streamlit>=1.24.0
together>=0.2.0
stripe>=7.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
pandas>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pypdf>=3.0.0
requests>=2.31.0
aiohttp>=3.8.0